+++ tests/shared/hardware_utils.py
@@ -3,7 +3,7 @@ import csv
 import logging
 from builtins import range
 
-from fonduer.learning.utils import entity_confusion_matrix
+from fonduer.learning.utils import confusion_matrix
 from fonduer.supervision.models import GoldLabel, GoldLabelKey
 
 try:
@@ -160,7 +160,7 @@ def entity_level_f1(
             else:
                 entities.add((doc, p))
 
-    (TP_set, FP_set, FN_set) = entity_confusion_matrix(entities, gold_set)
+    (TP_set, FP_set, FN_set) = confusion_matrix(entities, gold_set)
     TP = len(TP_set)
     FP = len(FP_set)
     FN = len(FN_set)