+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Added
 
 Changed
 ^^^^^^^
+* `@HiromuHota`_: Incorporate entity_confusion_matrix as a first-class citizen and rename it to confusion_matrix because it can be used both entity-level and mention-level.
 * `@HiromuHota`_: Separate Spacy#_split_sentences_by_char_limit to test itself.
 * `@HiromuHota`_: Refactor the custom sentence_boundary_detector for readability and efficiency.
 * `@HiromuHota`_: Remove a redundant argument, document, from Spacy#split_sentences.